requirements.txt
tensorflow-gpu==2.6.0
numpy==1.19.5
keras==2.6.0
pandas==1.1.5
efficientnet==1.1.1
opencv-python==4.2.0
Pillow==7.1.2
